requirements.txt
feedparser>=6.0
aiohttp>=3.9
newspaper4k>=0.9
scikit-learn>=1.3
requests>=2.31
anthropic>=0.40
python-dotenv>=1.0
